refactor: Move lock stuff in its own folder
